added controller file to clean up server.js, deleted config folder,  updated main scripts
