--- config/connection.js
@@ -1,28 +1,0 @@
-// Set up MySQL connection.
-var mysql = require("mysql");
-var connection;
-
-if(process.env.JAWSDB_URL){
-  connection = mysql.createConnection(process.env.JAWSDB_URL);
-}
-else {connection = mysql.createConnection({
-  port: 3306,
-  host: "localhost",
-  user: "root",
-  password: "Bk2=~%pqbV=ECC58",
-  database: "mongoheadlines"
-});
-};
-
-
-// Make connection.
-connection.connect(function(err) {
-  if (err) {
-    console.error("error connecting: " + err.stack);
-    return;
-  }
-  console.log("connected as id " + connection.threadId);
-});
-
-// Export connection for our ORM to use.
-module.exports = connection;
